Replace 'take action' button with address search field, gated via feature flag (#33)
This replaces the "take action" button with an address search field that, when submitted, sends the user to the data-driven onboarding results on app.justfix.nyc.

The search bar works with and without JS.  If the GeoSearch API isn't working, the field also reverts to its baseline experience (just a standard text field).

## Code reuse

Ideally we should be reusing `geo-autocomplete.tsx` and possibly even SCSS code from the tenant platform, rather than copy-pasting it, but since this is super high-priority I'm just copypasta'ing for now and deferring better code reuse to https://github.com/JustFixNYC/tenants2/issues/884.

## Feature flags!

Because I'd like to merge this into `master` ASAP without necessitating changes to our production site's front page, I've added a new [feature flag](https://en.wikipedia.org/wiki/Feature_toggle) in the form of an environment variable that controls whether or not to enable the new search bar, or to simply show the legacy "take action" button that's currently on the front page.  By default, the legacy behavior is preserved, but the new search bar can be enabled by setting `ENABLE_DDO=1` in one's `.env` file.  Once we are ready to launch, we can remove the feature flag and always display the search bar.