made all tenant app links on demo site go to demo.justfix (#93)
I made any hard-coded link to the tenant app go do demo.justfix.nyc when on the DEMO org site and app.justfix.nyc otherwise. Unfortunately, any links to the tenant app that we have in our Contentful space (like the links in the #products section, or in a Learning Center article) are not dynamic.