Add a changelog (#97)
This adds a changelog at `/changes` that is roughly based on the Slack changelog: there's a heading for each year and month (e.g. "October 2019") that has at least one change in it, and each individual change is listed with its own sub-heading title (e.g. "HP Action now in beta").  The entries are sorted reverse-chronologically.  When two entries happen to occur on the same date, they're sorted alphabetically.

Currently this page isn't linked-to from anywhere else; as such it's a kind of "soft launch" of the feature.