Refactored Mailchimp signups in footer to use new external API (#112)
This PR replaces our current API for Mailchimp email signups and instead uses the new API endpoint implemented in the Tenant Platform! It also adds some response messages below the email input to properly inform users of the status of their email signup. All is analogous to this PR on WhoOwnsWhat.