Make locale prefixes part of all routes (#124)
This PR makes locale prefixes mandatory for all page routes— as in, any English page will now have the locale prefix of /en/ within the URL. To make sure we don't break pre-exisiting links, this also introduces a new locale-redirect.tsx component that redirects any locale-unspecified routes to the default locale of the browser. For users missing javascript, we have a meta refresh tag to redirect them, as well as a message with a link to the otherwise-redirected route will appear after about 5 seconds of rendering the locale-redirect component.