Added a hook to grab the current page locale (#130)
This PR introduces a new hook called useCurrentLocale() that we can use to grab the current locale specified in the url of the current page. To try it out, this PR also refactors the <Header> component as well as our 404 page, so that they can make use of the new hook.

This PR also refactors the way we hold on to some of our global locale configurables, making a new locale-config.json file to store them in and exporting them where they are needed.