Internationalze default content. Fixes #133. (#136)
This fixes #133 by retrieving the default content for all locales and only using the one for the current page being rendered.

I'm guessing this might actually mean that JS-enabled browsers will also be downloading default content for all locales, too, but I've been unable to verify this via developer tools because I don't understand enough about how Gatsby is doing things. However, even if this is the case, I don't think it should be too big a deal because the metadata per locale isn't that big, and we currently only have two locales.