Fix homepage overflow issue caused by section divider (#180)
This PR fixes a bug where the content of the homepage extended beyond the <body> element on mobile, creating an ugly right margin. Turns out it was an issue with a section divider on the homepage. This refactoring addresses the issue by creating a new component that renders this section divider differently between desktop and mobile.