Add the Tools page to the org site (#213)
This PR implements the Tools/Products page based on Tahnee's designs in Figma. By implementing this page, this PR creates two new shared components of note:

a ReadMoreLink component that generates a link with icon when given a url. It also determines which link component and icon to use based on whether the url is an internal or external link
a PageHero component that creates the top-of-page full screen hero we are seeing on nearly every page but the homepage
This PR is also a good example of what is needed to create a new page (english and spanish) on the org site, which we will need to do again for the reports/analyses page.
a ReadMoreLink component that generates a link with icon when given a url. It also determines which link component and icon to use based on whether the url is an internal or external link
a PageHero component that creates the top-of-page full screen hero we are seeing on nearly every page but the homepage
This PR is also a good example of what is needed to create a new page (english and spanish) on the org site, which we will need to do again for the reports/analyses page.