Fix object reference issue when grabbing asset alt text (#219)
This quick PR fixes an error with our Netlify builds where we weren't thoroughly checking whether an asset's "description" field was null or not before assigning it to the asset's alt text. Sometimes folks will upload assets to Contentful without inputting the description field (and there's also no way to make that required), so this small change just fixes the logic of that check on our end so that our build doesn't break when someone forgets to include an description on a photo.