Update graphql queries for learning center pages (#231)
This PR updates our graphql queries to Contentful for all of our Learning Center pages. It also removes some old learning center components that are no longer used in the new Figma designs.

Note that because the pathnames for learning center pages are dynamically defined in Contentful, we can't rely on the convenience of our /pages directory to generate the article pages and instead need to generate these pages in gatsby-node.js. The one exception is the learning center homepage which can be found in pages/learn.en.tsx.