Add banner and modal notifying users of our cookies policy (#234)
This PR adds a banner that notifies user of our cookies policy and opens up a modal that explains more details about it.

NOTE: this new changes makes use of the react localStorage feature. In order to reset your local storage, you need to re-open localhost in an incognito window.