fix missing mobile photos, tidy spacing (#275)
This PR fixes an issue on mobile for the team page where the team member photos were hidden for what would be the middle columns on desktop. This was because of some slightly awkward nested columns required to get the team member cards to conform to the alignment in mocks. We have 3 size-4 columns, each with a single size-9 column inside that is either pulled-left, -right, or centered within their size-4 column. For some reason on mobile the centered ones had the images hidden, so in the scss file they are pulled left on mobile.

also tidy up some minor spacing things