use product name slug as element key (#289)
On the tools and homepage we show a list of all our product cards, and shuffle the order they are displayed each time. We noticed that on the first render they all are correct, but on refresh all the text of the cards is updated but the url for the button is unchanged. We believe the issue is that react is not re-rendering the button fully in production because we are using the product list array index as the element key. This PR changes to use the product name slug as the key instead.