remove unnecessary code from shuffle debugging (#305)
In previous PRs (#289, #290) we were trying to solve an issue of the shuffle order of product not correctly updating button links, but after discussing with the team we decided to remove the shuffle and go with a set order. This PR just removes some of that old code we don't need anymore.