tablet use desktop text styles and mobile layout(#312)
This PR does a first rough pass for tablet styles to serve as a baseline/reference for pairing with Tahnee tomorrow to refine. For tablet it uses the mobile layouts but the desktop text styles.  This PR builds off of #311 restoring the mobile layout for tablet, and using touch instead of mobile for classes