Use default order for funders, hide impact reports (#390)
We used to alphabetically sort our funders, but now will manually manage the order on contentful and use that default.

We haven't updated our impact reports recently, so will temporarily hide that section of our about/mission page until we have a more up-to-date version. (I just commented this out, but let me know if you think it's better to just delete and maybe just add a comment to this PR so it can easily be found to add back in.)